[FEAT] Build Error fix
