+++ src/component/tag/tag-filter-modal-selector.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 'use client';
 
 import CrossSvg from '@/asset/svg/cross/cross-svg';
